Extract reusable Core Data extensions into a separated file.
